Move core::probe::kernel mod.rs -> kernel.rs
